Split renderer/main.tsx — keyboard, menu router, repos store, tree expansion
